+++ extension/lib/mapper.js
@@ -6,8 +6,8 @@
  */
 module.exports.mapCountry = function (country) {
   const map = country && {
-    ...(country.length === 2 && {country_code: country}),
-    ...(country.length > 2 && {country})
+    ...(country.length === 2 && { country_code: country }),
+    ...(country.length > 2 && { country })
   }
   return map || {}
 }
@@ -18,8 +18,8 @@ module.exports.mapCountry = function (country) {
  */
 module.exports.mapProvince = function (province) {
   const map = province && {
-    ...(province.length === 2 && {province_code: province}),
-    ...(province.length > 2 && {province: province})
+    ...(province.length === 2 && { province_code: province }),
+    ...(province.length > 2 && { province: province })
   }
   return map || {}
 }
@@ -32,8 +32,8 @@ module.exports.mapProvince = function (province) {
  */
 module.exports.mapCustomAttributes = function (customAttributes) {
   const map = customAttributes && {
-    ...(customAttributes.hasOwnProperty('company') && {company: customAttributes.company}),
-    ...(customAttributes.hasOwnProperty('phone') && {phone: customAttributes.phone})
+    ...(customAttributes.hasOwnProperty('company') && { company: customAttributes.company }),
+    ...(customAttributes.hasOwnProperty('phone') && { phone: customAttributes.phone })
   }
   return map || {}
 }