+++ extension/user/addAddress_v1.js
@@ -1,8 +1,7 @@
 const Tools = require('../lib/tools')
 const UnauthorizedError = require('../models/Errors/UnauthorizedError')
-const InvalidCallError = require('../models/Errors/InvalidCallError')
 const SGShopifyApi = require('../lib/shopify.api.class')
-const {mapCountry, mapCustomAttributes} = require('../lib/mapper')
+const { mapCountry, mapCustomAttributes } = require('../lib/mapper')
 
 /**
  * @param {SDKContext} context
@@ -13,25 +12,20 @@ module.exports = async function (context, input) {
     throw new UnauthorizedError('User is not logged in.')
   }
 
-  if (Tools.isEmpty(input.address)) {
-    throw new InvalidCallError('Empty address data.')
-  }
-
   // Map the input address values to fit the Shopify specifications for the admin api endpoint
-  const address = Object.assign(input.address)
   /** @var {ShopifyAddress} newAddress */
   const newAddress = {
-    address1: address.street1,
-    address2: address.street2,
-    city: address.city,
-    first_name: address.firstName,
-    last_name: address.lastName,
-    province_code: address.province,
-    zip: address.zipCode,
-    name: address.firstName + ' ' + address.lastName,
-    ...mapCountry(address.country),
-    ...mapCustomAttributes(address.customAttributes),
-    tags: address.tags
+    address1: input.street1,
+    address2: input.street2,
+    city: input.city,
+    first_name: input.firstName,
+    last_name: input.lastName,
+    province_code: input.province,
+    zip: input.zipCode,
+    name: input.firstName + ' ' + input.lastName,
+    ...mapCountry(input.country),
+    ...mapCustomAttributes(input.customAttributes),
+    tags: input.tags
   }
 
   return new SGShopifyApi(context).addAddress(context.meta.userId, newAddress, (!Tools.isEmpty(input.address.tags) && input.address.tags.includes('default')))