+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Fixed
+- error handling for Shopify API and validation errors
+
+## [1.2.3] - 2018-08-30
+### Fixed
+- Fix bug that user can't login if the mail address contains capital letter
 
 ## [1.2.2] - 2018-08-30
 ### Added
@@ -15,5 +21,6 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - user not logged in after registration in webcheckout
 
 [Unreleased]: https://stash.localdev.cc/projects/SGX/repos/shopify-user/compare/commits?targetBranch=refs%2Fheads%2Fmaster&sourceBranch=refs%2Ftags%2Fv1.2.2
+[1.2.3]: https://stash.localdev.cc/projects/SGX/repos/shopify-user/compare/commits?targetBranch=refs%2Ftags%2Fv1.1.5&sourceBranch=refs%2Ftags%2Fv1.2.3
 [1.2.2]: https://stash.localdev.cc/projects/SGX/repos/shopify-user/compare/commits?targetBranch=refs%2Ftags%2Fv1.1.5&sourceBranch=refs%2Ftags%2Fv1.2.2
 [1.1.5]: https://stash.localdev.cc/projects/SGX/repos/shopify-user/browse?at=refs%2Ftags%2Fv1.1.5